[BUGFIX] Use cHash in language menu links if necessary
Current menu implementation does not add the cHash in language
menus. This causes a non-cached page and performance
degradation if the page contains parameters and a user switches
between languages. Together with extensions like RealURL it
can become a long time issue because of the caching of wrong
URLs. The fix implements analysis for the query string and
relevant parameters to decide whether the cHash should be
added or not.

Change-Id: I711e732555e03c9bf5a1321523b505735aa20c41
Resolves: #33833
Releases: 6.2
Reviewed-on: https://review.typo3.org/21242
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-by: Jigal van Hemert
Reviewed-by: Georg Ringer
Tested-by: Georg Ringer