[TASK] Move TCA and ext_tables.sql from ext:cms to ext:frontend
EXT:frontend is the successor of ext:cms. This started with the
namespace change so all class files are already located there.
The patch moves TCA and ext_tables.sql to the extension to make
it very clear that functional tests that depend on this data
should also go to ext:frontend later on.
The patch is trivial since both extensions are "required" and
thus always loaded, with ext:frontend directly before ext:cms,
so nothing can break with this patch.

Change-Id: Ifd9e0c9263f6b9a6aadfbfa1b568c4cd085e40d9
Resolves: #51251
Releases: 6.2
Reviewed-on: https://review.typo3.org/23244
Reviewed-by: Stefan Neufeind
Tested-by: Stefan Neufeind
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring