[BUGFIX] Clean-up responsibilities of FAL classes
This is a followup for the patch
I6ceeb8db4dd4e470e7c72dcfbc2121093411da75.

The latest change sets missed to adapt
the unit-tests accordingly.

Furthermore due to this change it came to light,
that the active loading option of storages in the file object
may create a recursion issue, if the file is not indexed.
This architectural problem has been removed by making
the constructor argument mandatory and extend the
ResourceFactory to always inject the correct storage.

Releases: 6.2
Resolves: #51528
Change-Id: Ic1561b83f49783e3af1b8f3403e1b8eab84a1380
Reviewed-on: https://review.typo3.org/24623
Reviewed-by: Frans Saris
Tested-by: Frans Saris
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn