[TASK] Make section index from hidden headers work again
Before version 4.7 you could build up a section index with
links to content elements with hidden headers.
Since version 4.7 elements with a hidden header type are sorted
out for the section index, overriding the extra flag "sectionIndex",
which is especially integrated for this.

This patch adds a extra flag for menu configuration sectionIndex
named "includeHiddenHeaders" to make it possible to re-enable
the behaviour of versions before 4.7.

Resolves: #52928
Related: #31767
Documentation: #53699
Releases: 6.2
Change-Id: Iee7f20cfc317c2b624dfc440811c5edc0a367a1e
Reviewed-on: https://review.typo3.org/24876
Reviewed-by: Markus Klein
Tested-by: Markus Klein
Reviewed-by: Wouter Wolters
Reviewed-by: Sebastian Fischer
Reviewed-by: Jigal van Hemert
Tested-by: Jigal van Hemert