[TASK] Fix remaining calls to FileRepository for retrieving Files
In previous efforts the calls to the FileRepository, when retrieving
File objects have been relocated to the ResourceFactory, since the
coexistence causes problems when it comes to metadata and multiple
instances of the same entity.

As findByUid is defined in the AbstractRepository it was overlooked
until now. This patch overrides the base method, deprecated it and
calls the ResourceFactory, too.
In addition all calls in core are replaced with the direct call.

Resolves: #53690
Releases: 6.2
Change-Id: I89a665211bd7d400f339e27bd2b155958fe5daf1
Reviewed-on: https://review.typo3.org/26850
Reviewed-by: Ingo Schmitt
Tested-by: Ingo Schmitt
Reviewed-by: Frans Saris
Tested-by: Frans Saris