[BUGFIX] Adding new menu types triggers error
With the addition of categories-related menus, new fields were added
to the tt_content "menu" palette for handling categories relations.
This causes problems when people define new menu types.

This patch adds the categories-related fields in a more targeted
way using subtypes_addlist. The difference is that these fields
are not *in* the palette anymore, but just after it. I think
it is okay from a usability point of view and it makes the whole
categories thing less intrusive.

Resolves: #54041
Releases: 6.2
Change-Id: I923a6ec0d7716821626a31a95d80027987038bf1
Reviewed-on: https://review.typo3.org/27591
Tested-by: Dirk Klimpel
Reviewed-by: Dirk Klimpel
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters