[BUGFIX] Using files in eID results in fatal error
In TYPO3_MODE == 'FE' we attached language overlays for file metadata
via an Signal in the MetaDataRepository. Within this signal the
methods of the PageRepository in $GLOBALS['TSFE'] are used to do the
actual language overlay.

When using eID the Frontend is not initialized. When creating file
objects in eID these functions are accessed on non-objects which
obviously leads to a fatal error.

As a result the registration of the Slot is extended to not register
itself in case of eID requests.

Resolves: #54232
Releases: 6.2
Change-Id: I4ccf226f48eb85428ce281c4a941ff7e834c2f88
Reviewed-on: https://review.typo3.org/27692
Reviewed-by: Wouter Wolters
Reviewed-by: Frans Saris
Tested-by: Frans Saris
Reviewed-by: Georg Ringer
Tested-by: Georg Ringer