[BUGFIX] Adapt ShowPic popup to use FAL
If not using own Lightboxes TYPO3 ships a custom solution
of an image-enlarge popup. This combines an eID for rendering
the image as well as a short JavaScript snippet opening that
in a new window.

The php code for rendering that link as well as the popup code
itself are not fal aware and to image-processing on their own.
As a result this feature is only able to deal with files in local
file system.

This change refactors both imageLinkWrap as well as
ShowImageController to use and work on FAL objects.

Releases: 6.2
Resolves: #56723
Change-Id: Iebf1379b8f72cf57d69a06b0b7540cbd8ccb32c1
Reviewed-on: https://review.typo3.org/28252
Reviewed-by: Frans Saris
Tested-by: Frans Saris
Reviewed-by: Fabien Udriot
Tested-by: Fabien Udriot
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter