[BUGFIX] Tests: Incomplete mocking in ext:frontend
Improve the fluid standalone tests by not calling __construct() of
a dependency.
Additionally, renderWorksWithNestedFluidtemplate() not only tests
the standalone view, but also its connection to the content object
renderer, so this is more a functional than a unit test. The test
is streamlined and moved to the functional test suite.

Change-Id: I3ae357de4861395622bc6f70aeed6011c7760cde
Resolves: #56788
Related: #51436
Releases: 6.2
Reviewed-on: https://review.typo3.org/28274
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn