[BUGFIX] Failing group/select test changeElementSorting
The workspace modify test "changeElementSorting" is failing
in the frontend for the TCA types "select" and "group".

The reason is, that the move placeholder is not selected
via the "uidInList" TypoScript property and thus removed
in PageRepository::versionOL().

To avoid that, "uidInList" gets extended to look-up
move placeholders if versioning preview is enabled.

Resolves: #56782
Releases: 6.2
Change-Id: Ifae30527bead02b38888f326de2615f6bfe35f5b
Reviewed-on: https://review.typo3.org/28278
Reviewed-by: Georg Ringer
Tested-by: Georg Ringer
Reviewed-by: Thorsten Kahler
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn