[!!!][BUGFIX] Respect selected target for shortcut mode 'Parent page'
The selected page is now respected when selecting the parent page
to use. An upgrade wizard ensures that existing shortcuts will not
break by removing the currently selected page.

The 6.2 version of this patch will only adjust the label to not
mention the "selected page" part, so the label matches
the current behaviour.

Change-Id: I51bc9239871d93728e84a4fe2e3b8bc9bb4a7404
Resolves: #57089
Releases: master, 6.2
Reviewed-on: http://review.typo3.org/28542
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>