[BUGFIX] Respect config.debug setting for displaying parsetime
2 possible information snippets can be outputted using
config.debug = 1 in TypoScript or [FE][debug]=1 in Install Tool.

- Parsetime
- Cached page generated time + expire date

TypoScript should override the general setting of the Install Tool
to make it possible to turn debugging of for specific PAGE objects.

This is especially needed if using e.g JSON output is invalid
having this information.

The change uses the same check as in index_ts where it works perfectly
fine.

Change-Id: I1520896476d97286627ab7f3b232b733dd3c50f7
Resolves: #57365
Releases: 6.2
Reviewed-on: https://review.typo3.org/28909
Reviewed-by: Helmut Hummel
Reviewed-by: Stefan Neufeind
Tested-by: Stefan Neufeind