[BUGFIX] Invalid localization overlay of cObject FILES
The TypoScript cObject FILES does not consider localization
overlays and behavior correctly when it comes additional 
translation table (e.g. for pages).

Concerning pages_language_overlay.media, there is no default
fallback to the values of the default language. This behavior
is the same to TYPO3 4.5. If the localization shall fallback
to the default language, the "mergeIfNotBlank" property needs
to be used for the accordant field.

For having the localization handling working properly the
FILES cObject section "references" must not set the properties
"table" and "uid" - otherwise the values will just be overriden
by those TypoScript settings.

Resolves: #58352
Resolves: #47869
Releases: 6.2
Change-Id: I8e915654e5ab1ed1eafe86411930f9fbfcc225ae
Reviewed-on: https://review.typo3.org/29739
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter