[BUGFIX] Complete datatable structure fe_groups
While creating frontend groups via extbase without any subgroup set,
the record will have the value NULL in the database
for the property subgroup.
This leads to an exception.

This patch only defines TSconfig, subgroups and description not
to accept NULL and to be consistent the tx_extbase_type gets
the default value 0.

Resolves: #60245
Releases: 6.3, 6.2
Change-Id: Ie402d4c47058755efe40cc762f4220ba7dc6b781
Reviewed-on: http://review.typo3.org/31560
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Susanne Moog <typo3@susannemoog.de>
Tested-by: Susanne Moog <typo3@susannemoog.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>