[TASK] Split up typolink() functionality in multiple methods - Part 1
The cObj->typolink() method is one of the biggest code chunks in
the CMS core, and one of the most central ones for the frontend.
However, it is hard to test, debug and understand. That's why certain
separate parts can be split up in multiple methods.

The first step provides the following changes:

* More readable variables within the method
* Split up the detection, whether it is a mail link,
  external url etc in a separate method
* Split up the string explode functionality of "23 - myclass"
  into a separate method

There are unit tests missing. Also, resolving each link type
could be put in a separate method as well in a further step.

Resolves: #49575
Releases: master
Change-Id: I40fe9367d65d6f6b7a857f5972765d1c2eb561ca
Reviewed-on: http://review.typo3.org/21785
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Tested-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Reviewed-by: Alexander Opitz <opitz.alexander@googlemail.com>
Tested-by: Alexander Opitz <opitz.alexander@googlemail.com>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>