[BUGFIX] Respect HTTP headers from JsonView
When rendering the json view, headers are set to the response
which are then sent by Extbase after the plugin is rendered.

However this takes place before the core (TypoScriptFrontendController)
sends the same header unless deactivated.

This change adds a new property to TypoScriptFrontendController to
be able for plugins to set the content type header to be sent.

Additionally the JsonView now makes use of this.

This is a temporary solution unless we have a proper
request/response model for frontend rendering.

Releases: master
Resolves: #60624
Change-Id: I29fa92e21495864a3c7777bbb0bc8e375267850e
Reviewed-on: http://review.typo3.org/34839
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>