[BUGFIX] Fix image rendering for renderMethod = table
Since the introduction of FAL to TYPO3, the rendering of images
using CSS styled content with renderMethod = table is broken.

Using a single image results in the image being rendered twice
and if you have multiple images in one content object, only the
last image is rendered multiple times (depending on the amount
of images in the content object).

This patch resolves both issues and loads the current image file
reference to the contentObjectRenderer while looping through the
images of the content object.

Resolves: #46864
Related: #47004
Releases: master, 6.2
Change-Id: If120267111e30de6438cecc5e14cea34bc6fa235
Reviewed-on: http://review.typo3.org/35407
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>