[TASK] Deprecate thumbnail functionality via thumbs.php
With FAL and the generation of preview images
in the backend via FAL, thumbs.php and its
controller are no longer in use.

The corresponding code can therefore safely
be marked for removal.

The only place still in use was filelist() in
the frontend, where FAL is now used for
rendering icon / preview images.

Resolves: #64068
Releases: master
Change-Id: Ice806fc1684fddc303468878e351866e52a5d8e4
Reviewed-on: http://review.typo3.org/35694
Reviewed-by: Frans Saris <franssaris@gmail.com>
Tested-by: Frans Saris <franssaris@gmail.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>