[BUGFIX] Allow page.config.pageTitleSeparator
The TypoScript option page.config.pageTitleSeparator
does not work currently, however the unspecific option
config.pageTitleSeparator does work.

This is due to the fact that the PageGenerator class
is preparing all options with TSFE->config and the
TemplateService class is putting everything together
with tmpl->setup which is not the consolidated config
data for this page. In order to streamline the code,
the pageTitleSeparator logic is moved
to PageGenerator where TSFE->config is used instead.

Resolves: #54410
Releases: master, 6.2
Change-Id: I677a852f6a94ea9737dade6e32e19f8b86dc9c89
Reviewed-on: http://review.typo3.org/35750
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>