[FEATURE] Add PageRepository::getPagesOverlay()
PageRepository::getMenu() calls getPageOverlay() for each single page
in a menu, causing a lot of queries to the SQL database.

This patch introduces getPagesOverlay() which is able to fetch multiple
page record overlays at once.
getMenu() is changed to make use of the new method.

It reduces the number of SQL requests on an uncached "features" page
from the introduction package from 721 to 648.

Change-Id: I014c08b6b63bcee99f94567240f1835077f60f0b
Resolves: #64258
Releases: master
Reviewed-on: http://review.typo3.org/36054
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>