[TASK] Don't define extTarget = _top as default
In the past, External links not generated via RTE/parseFunc,
e.g. just plain typolink, had the target="_top" attribute, which
in sites without frames is the same as not setting the target.

As frames are extraordinary these days, the default should be
to not set the extTarget at all by default, but let CSS Styled
Content set the extTarget to blank.

Resolves: #65783
Releases: master
Change-Id: Ifba46110ad4d099691530421c34b54e9da502f69
Reviewed-on: http://review.typo3.org/37897
Reviewed-by: Thomas Schlumberger <thomas@b13.de>
Tested-by: Thomas Schlumberger <thomas@b13.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>