[TASK] Move new content element wizard configuration to css_styled_content
The configuration for the new content element wizard is defined in the
extension frontend. The content elements which are defined are specifically
to css_styled_content. For the new frontend rendering some content elements
will not be used anymore.

This patch moves this configuration from frontend to css_styled_content
in a seperated file

Resolves: #65901
Releases: master

Change-Id: Idea403c64775ccdf1dd792229e594a470ad72ff9
Reviewed-on: http://review.typo3.org/38007
Reviewed-by: Oliver Eglseder <oliver.eglseder@in2code.de>
Tested-by: Oliver Eglseder <oliver.eglseder@in2code.de>
Reviewed-by: Markus Sommer <markussom@posteo.de>
Tested-by: Markus Sommer <markussom@posteo.de>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>