[FEATURE] Add SVG support
Added rendering support for SVG images. When a SVG image is scaled
there is no processed file created but only a sys_file_processedfile
record with the calculated new dimensions.

When a mask or explicit cropping is set for an SVG image a
processed file is created like for all other images.

An extra fallback is added to ImageInfo to determine SVG dimensions when
IM/GM fails.

Resolves: #50136
Releases: master
Change-Id: I9d4e298f0329059e83783a94209c80558ed83e27
Reviewed-on: http://review.typo3.org/38422
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Reviewed-by: Benjamin Kott <info@bk2k.info>
Tested-by: Benjamin Kott <info@bk2k.info>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>