[FEATURE] Introduce Request/Response based on PSR-7
The PSR-7 standard is adapted into the TYPO3 Bootstrap with a
backwards-compatible layer.

The PSR-7 implementation brings several new classes:
 * Message (the base for Requests and Responses)
 * Request (for Requests made within PHP)
 * ServerRequest and a factory based on the current system environment
 * Response
 * Uri (a unified API for fetching several parts of an URI)

At any TYPO3 request a new ServerRequest object is created inside the
Bootstrap and handed over to the RequestHandler which can then use this
object for checking certain GET and POST variables instead of using
GeneralUtility.

The proper call (usually a Controller) creates a Response object that
is handed back to the RequestHandler + Bootstrap. The TYPO3 Bootstrap
will output anything related in the shutdown() method.

An example is shown with the LoginController and currently hard-wired
as no proper routing/dispatching is there yet.

Currently this is an internal API as the rest (Dispatch/Router and
Controller API) will follow once the base is in.

Please note that the PSR-7 standard works with Value Objects meaning
that it is not possible to modify any object but instead new objects
will be created for Message, ServerRequest and Response if modified.

The next steps are:
* Integrate proper Routing + Dispatching for Backend Routes to register
  new BE requests
* Migrate all AJAX Calls to use the new API and request / response
  handling
* Introduce a common Base Controller for all regular BE requests which
  is based on Request/Response and works as a replacement for sc_base
* Then: proper documentation for the whole bootstrap /
  dispatch + routing / controller logic
* Integrate symfony console app into the CLI Bootstrap as alternative
  for Request/Response
* Refactor TSFE to use Response / Request objects properly
* Refactor redirects logic to use Response objects

Resolves: #67558
Releases: master
Change-Id: I5b528284ecca790f784c7780b008356158343ee8
Reviewed-on: http://review.typo3.org/40355
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>