[BUGFIX] Fix translated shortcut target in menus
This ensures that the target of translated "shortcut" pages is
properly available. Normally this value is lost due to language
overlay thus the original page record needs to be retrieved to
restore the original shortcut target.

Now it is possible to e.g. mark shortcut menu items as active
on translated pages.

Resolves: #67047
Related: #36822
Releases: master, 9.5
Change-Id: I2cabc252ba1936763d6cbe1fe0a093aa74f5a4e1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62477
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Susanne Moog <look@susi.dev>
Reviewed-by: Susanne Moog <look@susi.dev>