[FEATURE] Integrate typeahead.js for LiveSearch
The change removes ExtJS LiveSearch and introduces typeahead.js
as an AMD module alternative. The ExtDirect connector is thus
removed and a regular AJAX handler is added.

The special live search commands (the special treatments starting with #)
are put in the correct extensions where the DB table is set up.

Resolves: #67580
Releases: master
Change-Id: I3f5473164297b2d9121179ffd019af10caec821a
Reviewed-on: http://review.typo3.org/40419
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Andreas Wolf <andreas.wolf@typo3.org>
Tested-by: Andreas Wolf <andreas.wolf@typo3.org>