[FEATURE] Introduce "Application" classes for all entry points
In an effort to move away from TYPO3_MODE and global code
an instance that acts as an entry point has been introduced. The
Application class which serves as a wrapper for setting up the
Bootstrap and calling the request has been introduced.

Additionally, new Applications (like typo3_console) can benefit
from the flexible way of registering RequestHandlers on their own
while still using the bootstrap or only parts of it.

Resolves: #67808
Releases: master
Change-Id: I20c92d0bc8df5d50d76922a8dfe1cbd898e6f1dd
Reviewed-on: http://review.typo3.org/40493
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>