[!!!][TASK] Redesign FluidTemplateDataProcessorInterface
The FluidTemplateDataProcessorInterface introduced with #66907 has been
refactored to DataProcessorInterface.

This decouples it from the Fluid StandaloneView and makes the
ContentObjectRenderer available in the process method so the different
DataProcessor classes do no have to initiate it on their own.

Instead of manipulating the $data property of the ContentObjectRenderer
a new key/value store can be filled/manipulated by the different
dataProcessor classes.

Resolves: #67890
Releases: master
Change-Id: I461dbff039974715e2c8f916efd5d79c159cc8b5
Reviewed-on: http://review.typo3.org/40915
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>