[!!!][TASK] Cleanup dual-use of auth_timeout_field in AbstractUserAuthentication
So far the variable could either contain the name of a field or a
timeout-value in seconds.
Introduce a clean integer-field for the timeout and keep the current field
only for the case where a class wants to really fetch the timeout from a
field in the userdata.

Change-Id: I2751e06a315936119478d0dd81cc4c5b1e941368
Resolves: #68890
Releases: master
Reviewed-on: https://review.typo3.org/42464
Reviewed-by: Thomas Mayer <thomas.mayer@2bis10.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Daniel Goerz <ervaude@gmail.com>
Tested-by: Daniel Goerz <ervaude@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>