[BUGFIX] DataHandler: Assign reference default values
In MySQL stric mode scenarios, relation types not using MM tables
or the foreign_field definition (IRRE) might end up in no
references at all, which is only the trigger for the actual
problem. These references result in a blank string ("") that
is casted in non-strict modes to the accordant field type
defined by the DMBS schema.

In strict mode this has to be handled more specific. Since the
system cannot determine the real meaning automatically, values
for reference that are produced in the mentioned scenario are
now casted to zero (0) or to the default value defined in the
accordant TCA configuration of the field.

Resolves: #69004
Releases: master
Change-Id: I6608f1fe0f63a99caa1208b1d3d76354b6f659cf
Reviewed-on: http://review.typo3.org/42550
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>