[!!!][TASK] Remove charset functionality for locales
The TYPO3 frontend resolves
config.locale_all (e.g. set to de_AT.UTF-8) and stores
it in $TSFE->localeCharset by using a "best guess" based
on decade-old mappings.

$TSFE->localeCharset is only used in stdWrap.strftime
to convert the result from the localeCharset to the ->renderCharset.

However, as it is obvious that a misconfiguration of the system, which happens only when
config.set_locale = de_AT@iso-8859-15
config.renderCharset = utf-8
is set, this exception needs to be configured manually
when using stdWrap.strftime.charset = iso-8859-15

Resolves: #72826
Releases: master
Change-Id: I1bba231879ebaf8e8700099bb87a03aba5d1b562
Reviewed-on: https://review.typo3.org/46011
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>
Tested-by: Frank Nägler <frank.naegler@typo3.org>