[FEATURE] Add LinkService and "t3://" linking syntax
A new and unified syntax to link to pages, files, folders
within a TYPO3 instance is added. All links are prefixed
with "t3://".

This patch adds a LinkService to convert from the "old"
typolink syntax to the new syntax, and also allows to add
the new syntax to typolink syntax.

The further steps are
* Don't replace the <a> tags with <links-tags inside the RTE transformations anymore
* Make the frontend lib.parseFunc_RTE deal with <a> tags directly
* Add a migration wizard for existing content

Resolves: #74365
Releases: master
Change-Id: I95a68c5770f4e4c5d9f8b3973817f1a270353217
Reviewed-on: https://review.typo3.org/47094
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Bamboo TYPO3com <info@typo3.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>