[!!!][TASK] Remove entryPointPath from main applications
The entry point path is only used to calculate the number
of levels nested under the PATH_site so the PATH_site
constant can be calculated.

In order to be more flexible with paths, the hard-coded name
"typo3/" is removed and the calculation is based on the
number of subdirectories instead of naming the directories
which are irrelevant.

Resolves: #75323
Releases: master
Change-Id: Ibbad9ed803aaf588d540d19e0c7cd1bdce88d473
Reviewed-on: https://review.typo3.org/47378
Reviewed-by: Susanne Moog <typo3@susannemoog.de>
Tested-by: Susanne Moog <typo3@susannemoog.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>