[BUGFIX] Detect the conversion to string "0" in stdWrap_csConv
Check for the return value of the converter to be of type string or
return the input. The string "0" is detected.

If the return string of the converter is the empty string, return the
input. This is done to keep the previous behaviour apart from the "0"
issue.

Update the method comment.

Resolves: #76413
Releases: master
Change-Id: I1dc214a6b7cad5815d96c9cdb671862ee99b3063
Reviewed-on: https://review.typo3.org/48425
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>