[BUGFIX] Use context specific table in PageRepository
By invoking the ConnectionPool instance, it's possible to use
different database connections and different database platforms
for a particular table. If for instance the system is running on
MySQL in general, but one single table (not pages) uses SQLite,
PageRepository::enableFields() is using quoting characters for
MySQL in the SQLite context - which fails.

Resolves: #78301
Releases: master
Change-Id: I8317a22126b1725a1730ebc66042dbdbfc18fd7c
Reviewed-on: https://review.typo3.org/50231
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>