[TASK] Deprecate TypoScript option config.mainScript
The TypoScript option config.mainScript allows to set the frontend
entrypoint from "index.php" to something else.

As this has some bigger conceptual flaws (e.g. if this option is
set, backend preview will not work, as the backend preview does not
know about the selected TypoScript in the frontend), the option
is removed in favor of creating custom extensions to e.g. hook
into the frontend rendering to replace index.php.

Resolves: #78668
Releases: master
Change-Id: Idecdbc93fba33d9b8fecfa39c064fbeea4986a8d
Reviewed-on: https://review.typo3.org/50603
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Frederic Gaus <frederic.gaus@flagbit.de>
Tested-by: Frederic Gaus <frederic.gaus@flagbit.de>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>