[TASK] Restructure typolink tag generation
Information for creating the target and the class information etc. is
spread through the main function. Putting everything in one place and
using an array for building the <a> tag seems to be more appropriate.

The functionality for building the tag could now be moved in to a
separate function in a next step. Alternative would be to move the
resolving of the "target" field into a separate method.

Resolves: #80586
Releases: master
Change-Id: I38d33b64d75abbae2ed541e16c44fac66ab03f2e
Reviewed-on: https://review.typo3.org/51076
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>