[FEATURE] Introduce l10n_source field for tt_content
A new field which is a ctrl-based field for any TCA table
called "translationSource" set to a DB field name
(used for the new field tt_content.l10n_source) is added
and filled as it contains the uid of the record of where a
translation record came from.

This is different to "transOrigPointerField" as it
contains not the record to a record in the default language
or the one where a record was copied from, but rather
the information what the translation source was made.

Releases: master
Resolves: #78169
Change-Id: I3ecfeedb656c64681e5e3021cad6b708482520dc
Reviewed-on: https://review.typo3.org/51070
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>