[BUGFIX] Truncate long GIFBUILDER file names
GIFBUILDER generates file names from serialized data from
the GIFBUILDER setup (typoscript).
This can result in too long file names when the GIFBUILDER
is called with many parameters and on filesystems with
a max filename/path length.

The comment of the function fileName() states that the filename
should be shortened to 100 chars, but this doesn't happen.

This patch introduces a simple substring to enforce this limit.

Change-Id: Iaa17760bbf34fc6c01a9bc5ea8d1f8b6b242fecb
Resolves: #79301
Releases: master, 7.6
Reviewed-on: https://review.typo3.org/51305
Reviewed-by: Marco Huber <mail@marco-huber.de>
Tested-by: Marco Huber <mail@marco-huber.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>