[BUGFIX] Fix functional tests for EXT:extbase on PostgreSQL
Fix ExpressionBuilder::inSet() to allow working with columns
on a seondary table on Non-MySQL databases. Fix the usage of
inSet() within Typo3DbQueryParser to properly quote the table
identifier and ensure that all places that use inSet() use it
to compare a string value or a field on a table.

Fix assertions to take string/int differences into account and
ensure consistent order when asserting lists of records.

Change-Id: I9fecd228cde1a7d5a6202702616d806b4354eb4f
Resolves: #79796
Releases: master
Reviewed-on: https://review.typo3.org/51733
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Mona Muzaffar <mona.muzaffar@gmx.de>
Tested-by: Mona Muzaffar <mona.muzaffar@gmx.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>