[BUGFIX] Opening links in pop-up windows
Moved the check for if the target is coded as a JS open window link,
to a later stage in the function.

Since Css Styled Content and Fluid Styled Content was refactored,
the target might not be set until later in the typoLink
function. This caused pop links to stop working, moving the check for a
pop up link target to a later stages fixes the problem.

Resolves: #79918
Releases: master
Change-Id: Ie27fa3142df83c33ba3ad465ecb35c0f6e571f86
Reviewed-on: https://review.typo3.org/51830
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Anders Kostending <aha@systime.dk>
Reviewed-by: Faton Haliti <fha@systime.dk>
Reviewed-by: Jasmina Ließmann <code@frauliessmann.de>
Tested-by: Jasmina Ließmann <code@frauliessmann.de>
Reviewed-by: Timo Hund <timo.hund@yahoo.com>
Tested-by: Timo Hund <timo.hund@yahoo.com>
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>