[TASK] Deprecate cObj->URLqMark
The method cObject->URLqMark() adds a question mark to a string
to append it. This method should not be in scope of cObject, but rather
be put in e.g. StringUtility or GeneralUtility... however, since the functionality
is rather simple, it can be achieved without any helper method as a one-liner.

Resolves: #80510
Releases: master
Change-Id: I5230c5a4e27f8ef198190defd15598023279c1c2
Reviewed-on: https://review.typo3.org/52182
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>