[BUGFIX] Creating typolink in backend should not throw an exception
The global $TSFE object is not available in the Backend anymore.

In order to use <f:link.typolink> in e.g. a toolbar item or a preview
element, the TSFE object should be created as dummy so at least an index.php
URL (or external URL) is shown.

Resolves: #80669
Releases: master
Change-Id: I24a23a2d562c62e5663753bb80c75cc05164a6c9
Reviewed-on: https://review.typo3.org/52351
Reviewed-by: Frans Saris <franssaris@gmail.com>
Tested-by: Frans Saris <franssaris@gmail.com>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: TYPO3com <no-reply@typo3.com>
Tested-by: Mona Muzaffar <mona.muzaffar@gmx.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>