[BUGFIX] Allow arbitrary protocols in new LinkService
Since TYPO3 v8 it is not possible to simply create a link to "tel:+491727144385" anymore,
which was possible in TYPO3 v7. Instead an exception is thrown as the "toString()" method
cannot handle legacy links which have no specific type.

The patch checks for a "url" attribute allowing to represent a URL and to render a link via typolink.

Resolves: #80217
Releases: master, 8.7
Change-Id: I617e280295b4789a8c295f5ee21463a643e58afc
Reviewed-on: https://review.typo3.org/52684
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>