[BUGFIX] Remove wrong usage of PageRepository->error_getRootLine_failPid
The TypoScript frontend calls TSFE->sys_page->getRootline() which fills
the public property "error_getRootLine_failPid" when the third parameter
is set. Otherwise it throws an exception.

Setting this option was actually removed in TYPO3 when the RootlineUtility
was introduced, making the third parameter very non-useful, as it does not
throw exceptions but instead writing some weird values in the error_getRootLine
property.

It also goes back to the case where branch versioning was still supported.

In case of regular frontend requests there is a lot of dead code now because
of always-false checks on $TSFE->sys_page->error_getRootLine_failPid == -1.

Resolves: #81027
Releases: master, 8.7
Change-Id: I141f145c1855f09c33f003422c8e34653bf7f951
Reviewed-on: https://review.typo3.org/52621
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>