[TASK] Bootstrap: Load TCA configuration on every request
By moving Bootstrap::loadBaseTca() to the configure() call of the
TYPO3 Bootstrap, the global TCA variable is now populated very early
and it is obvious that this logic is then part of the actual system setup,
and does not need to be handled by any RequestHandler anymore.

Considering TCA is necessary for a TYPO3 environment to run,
this change might be very helpful.

Please note that the Install Tool application does not call
configure(), so this is not affected.

This also removes some issues related to Extbase
Command controllers in CLI handling.

Side effects:
- eID requests now have full TCA loaded as well at any time
- TCA is now loaded BEFORE entering any request handler check
- TCA is now loaded before e.g. BackendRouting initialization, and other Backend security checks

Easy to test the current behaviour:
- Clear all caches via install tool
- Call CLI command extbase:help which will result in a Warning / Exception

Resolves: #72222
Resolves: #81201
Releases: master
Change-Id: I31a0de0d0b039eb6180123678c61d3f26435dd63
Reviewed-on: https://review.typo3.org/52753
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Jan Helke <typo3@helke.de>
Tested-by: Jan Helke <typo3@helke.de>
Reviewed-by: Philipp Gampe <philipp.gampe@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>