[BUGFIX] Remove doubled SQL query in PageRepository
The method PageRepository->checkWorkspaceAccess() does a SQL query to fetch
the workspace record, then hands it over to BE_USER->checkWorkspace which has the
same functionality - to fetch the full workspace record.

As this is redundant code, the PageRepository code can be simplified a lot by just leaving
the workspaceCache property.

Resolves: #81187
Releases: master, 8.7
Change-Id: Ia82febb2269084ea9101acfc36a1c9aa20950290
Reviewed-on: https://review.typo3.org/52854
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>