[BUGFIX] Check for enabled devlog inside devLog()
Calling GeneralUtility::devLog() can be called but
the method itself should check whether the devlog
is enabled, instead of having third-party code
rely on checking the option with an if statement.

This effectively removes the need for checks on
if (TYPO3_DLOG) and the other two constants TYPO3_ERROR_DLOG
and TYPO3_EXCEPTION_DLOG in all callers' code.

Resolves: #80289
Releases: master, 8.7
Change-Id: Ibaeeca6300bdb37a513a07fb46c2616cfc51c7f2
Reviewed-on: https://review.typo3.org/52751
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>