[BUGFIX] AdminPanel: Change fields for simulate time
When entering a date string in the admin panel to simulate a certain
date/time and clicking the "update" button an exception it thrown. Due to
the names and ordering of the fields, the timestamp from the hidden field
gets overridden by the date string.

Move the hidden timestamp field after the human readable date string field
to avoid the timestamp being overridden by the date string. Add a comment
to explain the ordering and avoid future changes.

Resolves: #81174
Releases: master, 8.7
Change-Id: I87236c9b8ef84dabb3ec63e0530f14582ecbe6fe
Reviewed-on: https://review.typo3.org/53230
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>