[TASK] Remove TypoScript option config.typolinkCheckRootline
The TypoScript option 'config.typolinkCheckRootline' was introduced
back in TYPO3 v4.x when there was no proper rootline cache or sys_domain
cache in place, in order to resolve the proper domain.

This now has changed, and this option can be removed in order to simplify
page link generation.

When in a multi-site setup, this option MUST be active otherwise it will
result in ugly and hard-to-debug side-effects.

For a single-site setup having this option disabled will result in a marginal
performance decrease (one SQL query more) when linking to a page with a
sys_domain record available.

A deprecation notice is thrown if your TypoScript setup still includes this option.

Resolves: #82909
Releases: master
Change-Id: I36abcdd1e2eabe2da66d8edd720a5a27ec306ede
Reviewed-on: https://review.typo3.org/54545
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: TYPO3com <no-reply@typo3.com>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Frans Saris <franssaris@gmail.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>