[!!!][TASK] Remove config.lockFilePath functionality
When using the stdWrap.filelist functionality, the lockFilePath
only allowed to be used for a certain directory (usually, if not set, it
was set to fileadmin/), however, this is both very unsecure and inflexible.

Thus, stdWrap.filelist is extended to be used with all local FAL storages.

Add the same time, the public property TSFE->lockFilePath and the TypoScript
option "config.lockFilePath" is removed.

Resolves: #83256
Releases: master
Change-Id: Ia86c6686128dae4c0870cd15e019f4d53a4b28b6
Reviewed-on: https://review.typo3.org/54974
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Reviewed-by: Daniel Gorges <daniel.gorges@b13.de>
Tested-by: Daniel Gorges <daniel.gorges@b13.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>