[TASK] Separate Error handling from TSFE
The goal is to decouple error handling from TSFE.
For that, a new ErrorController is added which can be
called and returns a Response object instead of
exits the current request.

In order to keep full backwards-compatibility,
TSFE currently does not return a response, but
instead exists with a response object.

Next step:
- Move the pageUnavailable_force functionality into a middleware
- See what can be further abstracted to return a response

Resolves: #83883
Releases: master
Change-Id: I41f5f1a32a3f7edd48eb5c0464a26762ad4a946f
Reviewed-on: https://review.typo3.org/55690
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Tested-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>