[TASK] Use TSFE API methods for checking workspace preview
Instead of using the public "workspacePreview" parameter in TSFE,
which is only set when a sys_preview keyword is used, the parameter
from the BE User can be used instead directly (which is used anyways
when using the API methods TSFE->whichWorkspace()).

The TSFE method doWorkspacePreview() is adapted accordingly to be in line
with the rest of the code.

Resolves: #84038
Releases: master
Change-Id: Ic819e099abf36bda867bb2235d6aa23fbbfe19a4
Reviewed-on: https://review.typo3.org/55895
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Tested-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>