[TASK] Call getDomainStartPage without third parameter
The third parameter in PageRepository->getDomainStartPage
is not used anymore. Saving an extra call to GeneralUtility
to not hand this parameter over to the method in TSFE should be helpful.

However, in the future, this method will probably be changed,
the third parameter is not deprecated for now.

Resolves: #84043
Releases: master
Change-Id: I397dd86d703f4fa5778794209e90ffaa82baf872
Reviewed-on: https://review.typo3.org/55901
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Joerg Boesche <typo3@joergboesche.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>