[BUGFIX] Revert "[BUGFIX] Fix recursive FLUIDTEMPLATE cObj's that use layoutRootPaths"
This reverts commit 5c048a4c8017e4d50ab5f2343970a0fd33e4aaed
as introduced in https://review.typo3.org/54416

That workaround for a fluid bug had side effects when DataProviders
expected a ConfigurationManager instance to be available. That is
nothing officially supported but broke some (bad behaving) DataProviders.
As the fluid bug has been fixed in https://github.com/TYPO3/Fluid/pull/342
(contained in release 2.5.2) we can revert the workaround now.

The functional test that was introduced next to the bugfix is not reverted,
as do want to keep that obviously.

Resolves: #84071
Releases: master, 8.7
Change-Id: Ie6f5541f0f217fee7e85df954bb964a32fd27367
Reviewed-on: https://review.typo3.org/56035
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Manuel Selbach <manuel_selbach@yahoo.de>
Tested-by: Manuel Selbach <manuel_selbach@yahoo.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>